backend/app/requirements.txt
SQLAlchemy==1.3.13
starlette==0.13.2
uvicorn==0.11.3
fastapi==0.52.0
psycopg2==2.8.4
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-logging==1.15.0